requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
cryptography==41.0.4
deprecation==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
gotrue==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
nanoid==2.0.0
packaging==23.2
postgrest==0.11.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
realtime==1.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
typing_extensions==4.8.0
websockets==10.4
